Allocate only as much stack as is actually needed by functions. Also fix the max stack height (where applicable)

+++ tools/compiled/translator/instructions/catch.cc
@@ -69,7 +69,7 @@ public:
     return true;
   }
 
-  int getMaxStackHeight()
+  size_t getMaxStackHeight()
   {
     return 2;
   }
